Create README.md
This  is a caleder View to which we can modify easily based on our requirement. we need to use this code in a view controller where we need to use calender . 

Key features of this calender 
we can  easly differentiate the dates by different different colores .

How to use 
you need to store the dates in availableDates  Array in the project.




Note : Mail me (sandeepsiva512@gmail.com) if  any one have any doubts to use this calender.